backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
yfinance>=0.2.50
curl_cffi>=0.7.0
openai==1.51.0
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
httpx==0.27.2
pytz==2024.2
